--- Foogle-Search-Final/src/Page/Filterpage/FilterSearchPage.css
@@ -1,33 +1,0 @@
-.filter-container {
-    text-align: center;
-    margin-top: 100px;
-  }
-  
-  .logo {
-    width: 200px;
-    margin-bottom: 20px;
-  }
-  
-  .filter-options {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-bottom: 20px;
-  }
-  
-  .filter-label {
-    font-size: 18px;
-    margin: 5px;
-    cursor: pointer;
-  }
-  
-  .filter-button {
-    padding: 10px 20px;
-    font-size: 16px;
-    border: none;
-    border-radius: 20px;
-    cursor: pointer;
-    background-color: #34a853;
-    color: white;
-  }
-
