--- Foogle-Search-Final/src/Page/Searchpage/SearchPage.css
@@ -1,34 +1,0 @@
-.search-container {
-    text-align: center;
-    margin-top: 100px;
-  }
-  
-  .logo {
-    width: 200px;
-    margin-bottom: 20px;
-  }
-  
-  .search-input {
-    width: 50%;
-    padding: 10px;
-    font-size: 18px;
-    border: 1px solid #ccc;
-    border-radius: 20px;
-    outline: none;
-  }
-  
-  .button-container {
-    margin-top: 20px;
-  }
-  
-  .search-button {
-    padding: 10px 20px;
-    margin: 5px;
-    font-size: 16px;
-    border: none;
-    border-radius: 20px;
-    cursor: pointer;
-    background-color: #4285f4;
-    color: white;
-  }
-
