+++ foogle-discovery/README.md
