+++ foogle-discovery/public/logo512.png
