+++ foogle-discovery/public/manifest.json
