+++ foogle-discovery/src/Page/Home/Home.js
