+++ foogle-discovery/src/components/ImageGrid/ImageGrid.js
