+++ foogle-discovery/src/components/MainContent/MainContent.css
