+++ foogle-discovery/src/components/MainContent/MainContent.js
