+++ foogle-discovery/src/components/Navbar/Navbar.js
